Remove atomicModify tests
Here's the problem: atomicModify can't work, I don't think it ever did.
It relies on sole ownership of the redis connection for the duration of
the initial watch commands and follow up redis queries, but we cannot
guarantee it.

Also and relatedly: those tests cause flake failures in other tests.
Precisely because they can't own the connection while they run, the
`WATCH` commands they fire can be around when a test running in parallel
does a `MULTI .. EXEC`, and then fails with a 'watched key changed'
error.

Since there's no point in testing functionality we know doesn't work,
and test flakes are causing real pain, let's remove these tests.

We should also remove our use of the atomicModify command, because it's
making claims that it cannot live up to.