Modify lock configuration
One thing to consider when setting the amount of retries is how it will
relate to the max lock time. If the locked operation is allowed to take
a second but we only retry for 100ms, the retrying might not be super
useful.

This expresses both lock and retry time as a duration to make this more
explicit.