Ramp down sampling rate much more slowly
The average sample rate was much to low before, because we ramped down
too quickly as requests got slower. This commit makes it ramp down much
more slowly. It also simplifies the rampdown equation a bit.

It also sets the apdex back to 10ms, which is what we want our apdex to
be. We shouldn't change our apdex to fit our desired sampling rate,
there's other levers we can pull.