Standardize library test setup
In a previous commit we made `aide watch` run tests in libraries. To do
this it needs to watch the `spec` suite rather than the library itself.
For this to provide a nice experience that `spec` suite should know both
about the test and regular source code. This sets up our libraries to do
this all correctly.