Honeycomb sampling ramps down faster
Slow requests have a slightly higher likelyhood of being logged.

This is an ongoing attempt to optimize data into Honeycomb. We still
have a little bit of space to log additional requests so this should be
fine. If we want to favor slow requests even more we will likely need to
combine it with moving the base sampling rate lower to make room in our
Honeycomb budget.