Don't implicitly turn redis cmds into transactions
When we configure our handler to set a default redis expiry time, the
way we achieve this is by adding additional EXPIRE commands for all the
keys our redis commands touch. We used to add these EXPIRE commands to
an existing transaction, or create one if none existed yet. This commit
removes this automatic creation of redis transactions.

The reason for removing is that its implicit behavior that might go
against a conscious decision by programmers not to use a transaction.
For example: we might prefer sending separate commands knowing that
putting them in the transaction would create a huge and slow command
(this is a real case we've witnissed).

We don't gain much by putting the EXPIRE commands in a Redis
transaction. Redis transactions are a bit of a misnomer because they
don't give us transactionality. Commands won't be rolled back in case of
an error halway through. We do get some assurances because redis will
parse all commands and fail to evaluate any of them if that parsing step
fails, but we already have our Haskell API to prevent errors like that.

Redis 'transations' in our case only give us atomicity: commands in a
transaction will be evaluated one-after-the-other without other commands
being able to read or change data in between. We don't really need that
with expire's though, it's fine if the expiry of keys is updated a
command or two after the command the touched it.