Get span details in first pass
Before this commit the generation of span details happened in two steps.
First we generate noisy span details, then we remove the noise we don't
want in a second pass.

This makes it so we generate the non-noisy version in the first pass.
This makes it so we only need to consider a single place to figure out
why details look the way they do.