Change the Span type to basically be a dictionary
This is a refactor to prepare ground for an upcoming change. The
eventual goal is to rely less on the "details." namespace in honeycomb
keys, prefering a more specific namespace instead, i.e. "msql." for
mysql queries, "log." for log entries, etc.

Our old Span type isn't super supportive of this change. It makes the
assumption that there's some common fields and then a single dict of
details field that will all get the same prefix.

This commit makes it so we model our Span type a bit more like
honeycomb's own representation of a span: a dict of key value pairs
where the values can be any JSON object. This will allow us to pick the
keys we want.

As a side benefit this approach is that it should make our code a bit
more performant as well. We'll need to copy data around less, and can
rely on Aeson's efficient encoding strategy in more places.

Before we continue towards the goal though we'll need to address two
`Debug.todo`s this commit adds. If the `Span` type becomes opague we'll
no longer to easily read information out of it again. We'll need to grab
information from the original monitoring spans again. That's up next!