Allow network-uri 2.7.0.0
Version 2.7.0.0 of network-uri was retracted later and rereleased as
2.6.2.0, because all changes in it were backwards-compatible.

https://github.com/haskell/network-uri/issues/48

When setting the bounds to exclude it I assumed it was broken. I took
another look when https://packdeps.haskellers.com started reporting
nri-http wasn't supporting the latest version of network-uri.

It's probably best for folks not to use the version of network-uri
that's not recommended by the authors (it didn't bet minor and patch
level upgrades that were released on the 2.6 branch). 2.7.0.0 doesn't
seem _incompatible_ with nri-http though, so to make
packdeps.haskellers.com happy I suppose we can support it.