Remove MySQL code
We might open-source this as a separate library one day. It's a bit of
an odd one though, because we achieve our compile-time MySQL queries by
using a combination of compile-time Postgres support and a MySQL foreign
data wrapper. It's decidedly more experimental than our Postgres code,
so it probably makes sense to keep it separate.