Message ParamOffsets type
The core idea here is that we don't really need the topic name to be
part of the ParamOffsets type. The ParamOffsets type is used in the
context of setting up a subscription for a single topic, so the caller
already knows which topic we're talking about.

Dropping the topic name makes it so our fetch offsets function only
needs to provide the partition ids we're interested in, for which we can
use plain Ints. This allows us to specialize the `ParamOffsets` type for
representing a particular partition offset, and include an offset field.