Support All 1.x versions of time
time frequently has releases that bump the second version number, which
in PVP means a breaking change but doesn't usually seem to be breaking
for the time library.

In order not to have to perform these kinds of bump updates in the
future as often, we'll constrain our version of time to < 2.