Pass around the root `onFinish`
Replacing the parent does exactly what it implies: replaces the _direct_
parent.

In other words, it doesn't result in the behaviour of the root, but
rather in the behaviour of the direct parent. If an extra `Log.withContext`
is snuck in somewhere along the way, everything gets messed up again.

Having a way to directly mimick the behaviour of the root also allows us
to prevent the double-finish behaviour displayed by the final remaining
span.