Ensure a frozen callstack for generated tests
Before this commit, we effectively had this:

a :: Stack.HasCallStack => IO ()
a = do
  Stack.withFrozenCallStack b

b :: IO ()
b = do
  c

c :: Stack.HasCallStack => IO ()
c = do
  Stack.callStack & Stack.getCallStack & print

And the effect of this, is in `c`, we get a callstack with a single
frame: `b`'s call to `c`.

We need to have `HasCallStack` all the way through for
`Stack.withFrozenCallStack` to work.