Source setup-postgres.sh
Running it means the `export`s don't do anything, and our template
haskell can't figure which db to actually connect to, reverting to the
default. That means the behaviour depends on how those default
credentials work for the used version of postgres, and also just doesn't
seem to be the intent.

Sourcing the script means it is essentially inlined, as it used to be.